day 5/backend/requirements.txt
fastapi[all]>=0.110.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.30
langgraph>=0.0.34
chromadb>=0.4.24
python-dotenv>=1.0.1
httpx>=0.27.0
uvicorn>=0.29.0
langchain-groq>=0.1.0
langchain-openai>=0.1.0
